Few minor changes.
Floats are now treated as pixel values as well as Fixnums. Need to
refactor out that logic and make it configurable, ie. default to pixels
or percentages, or whatever. (You can override manually by using
strings but then you lose simple math capabilities.

Added indexing operators to CSSClass so you don't have to go through
the rules property, just to make the code a little less verbose.

Added default ease-in-out for the animation timing function. Will need
to change this to be configurable from the options parameter of
CSSAnimation#animation, but I'll get to that later as I'll probably
only want ease-in-out for the foreseeable future.