Moved samples, added organization, broke everything.
This commit breaks everything as it is not finished yet. How ever it is
more organized and ready to have actual work done to it.

Decided to implement the animations as a series of keyframe percentages
as opposed to a series of animations composing a single one. The
overhead of the JavaScript was too high and caused issues when trying
to play the same animation twice on two different elements.

This way is better, but will take more time to implement. But I'll get
there.