Updated syntax, helper methods, fleshed out CSSRule, added CSSKeyframe, no more JavaScript.
This is a big change on development. Updated syntax a bit.

* Rules should be added using the add_rule or add_rules method on
CSSClass so that the values can be wrapped in a CSSRule.

* JavaScript was dropped in favor of generating Keyframes with
percentage blocks. Much better performance and less chance of fucking
up generating JavaScript (which was more difficult)

* Multiple classes are now supported for a single animation so they can
be run at the same time to not only chain animations on single objects,
but animations on many at the same time.

* Moved a bunch of logic for generating CSS strings out of
CSSAnimation. Added CSSKeyframe and CSSKeyframeExporter to manage
keyframe states.

Still need to do some refactoring so that names of methods (especially
regarding generating CSS) are consistent.

Also need to regenerate the CSS animation classes.

Very close though.